+++ tests/src/Parsers/ObjectParser.test.js
@@ -1,5 +1,55 @@
-describe('Object Parser', () => {
-  test('works.', () => {
+const fs = require('fs');
 
-  })
+const FileParser = require('../../../src/Parsers/FileParser');
+
+lgd = {};
+// lgd.codeActions = CodeActions.create();
+// lgd.lgdDiagnosticCollection = vscode.languages.createDiagnosticCollection();
+// lgd.logger = Logger.create('LGD.FileParser');
+lgd.configuration = {
+  options: {
+    createDebugLog: false,
+    tabSize: 2,
+    extractPropsAndState: true
+  }
+};
+
+
+/**
+ * @description We shouldn't be trimming but there is a bug in jest that doesn't allow this check to work without trimming.
+ * @param {string} str
+ */
+function fixString(str) {
+  return str.trim();
+}
+
+function CheckFile(filePath) {
+  const originalFile = fs.readFileSync(`./tests/mocks/${filePath}.js`, 'utf8');
+
+  const fileParser = FileParser.create();
+  let parseResult = fileParser.parse('', originalFile);
+
+  const compiledFile = fs.readFileSync(`./tests/mocks/${filePath}.d.ts`, 'binary').toString();
+
+  // fs.writeFileSync(`./tests/debug/${filePath}.debug.d.ts`, parseResult.typeFile);
+
+  const typeFileAry = parseResult.split('\n');
+  const compileAry = compiledFile.split('\n');
+
+  expect(typeFileAry.length).toBe(compileAry.length);
+
+  for(let i = 0; i < typeFileAry.length; ++i) {
+    expect(fixString(typeFileAry[i])).toEqual(fixString(compileAry[i]));
+  }
+}
+
+describe('Class Parser.', () => {
+
+  test('Obj is being parsed when export is at front of obj.', () => {
+    CheckFile('ExportConst');
+  });
+
+  test('Template comments are parsed properly.', () => {
+    CheckFile('BaseCardView');
+  });
 });