+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to the "js-syntax-extension" extension will be documented in
   - You can extend classes besides React.Component now.
   - Adding props and state to React.Component<> by default.
 - export const obj = {}; is now being parsed.
+- Parsing template correctly if using comma. No longer have to define all templates on new lines.
 
 # 2.4.0+